+++ backend/app/Modules/Parties/Routes/api.php
@@ -5,22 +5,22 @@ use App\Modules\Parties\Controllers\Api\PartyTypeController;
 use Illuminate\Support\Facades\Route;
 
 Route::prefix('party-types')->group(function () {
-    Route::get('/', [PartyTypeController::class, 'index']);
-    Route::get('options', [PartyTypeController::class, 'options']);
-    Route::post('/', [PartyTypeController::class, 'store']);
-    Route::post('bulk-delete', [PartyTypeController::class, 'bulkDelete']);
-    Route::get('{partyType}', [PartyTypeController::class, 'show']);
-    Route::put('{partyType}', [PartyTypeController::class, 'update']);
-    Route::delete('{partyType}', [PartyTypeController::class, 'destroy']);
+    Route::get('/', [PartyTypeController::class, 'index'])->middleware('permission:party_type.view');
+    Route::get('options', [PartyTypeController::class, 'options'])->middleware('permission:party_type.view');
+    Route::post('/', [PartyTypeController::class, 'store'])->middleware('permission:party_type.view');
+    Route::post('bulk-delete', [PartyTypeController::class, 'bulkDelete'])->middleware('permission:party_type.delete');
+    Route::get('{partyType}', [PartyTypeController::class, 'show'])->middleware('permission:party_type.view');
+    Route::put('{partyType}', [PartyTypeController::class, 'update'])->middleware('permission:party_type.edit');
+    Route::delete('{partyType}', [PartyTypeController::class, 'destroy'])->middleware('permission:party_type.delete');
 });
 
 Route::prefix('parties')->group(function () {
-    Route::get('/', [PartyController::class, 'index']);
-    Route::get('source-options/{type}', [PartyController::class, 'sourceOptions']);
-    Route::post('/', [PartyController::class, 'store']);
-    Route::post('bulk', [PartyController::class, 'bulkStore']);
-    Route::post('bulk-delete', [PartyController::class, 'bulkDelete']);
-    Route::get('{party}', [PartyController::class, 'show']);
-    Route::put('{party}', [PartyController::class, 'update']);
-    Route::delete('{party}', [PartyController::class, 'destroy']);
+    Route::get('/', [PartyController::class, 'index'])->middleware('permission:party.view');
+    Route::get('source-options/{type}', [PartyController::class, 'sourceOptions'])->middleware('permission:party.view');
+    Route::post('/', [PartyController::class, 'store'])->middleware('permission:party.create');
+    Route::post('bulk', [PartyController::class, 'bulkStore'])->middleware('permission:party.create');
+    Route::post('bulk-delete', [PartyController::class, 'bulkDelete'])->middleware('permission:party.delete');
+    Route::get('{party}', [PartyController::class, 'show'])->middleware('permission:party.view');
+    Route::put('{party}', [PartyController::class, 'update'])->middleware('permission:party.edit');
+    Route::delete('{party}', [PartyController::class, 'destroy'])->middleware('permission:party.delete');
 });
